John conroy/refactor provenance (#1600)
* Remove redundant function and typography

* Remove unused styles

* Resolve prop type warning

* Pull derived entities into button component

* Simplify ShowDerivedEntitiesButton props

* Update test

* Factor out meta object in ProvData
