+++ context/app/static/js/components/Detail/provenance/ProvVis/__tests__/ProvData.spec.js
@@ -135,6 +135,7 @@ describe('CWL utils', () => {
       meta: {
         global: true,
         in_path: true,
+        type: 'data file',
       },
       name: 'name1',
       run_data: {
