+++ context/app/static/js/components/Detail/provenance/ShowDerivedEntitiesButton/index.js
@@ -1,0 +1,3 @@
+import ShowDerivedEntitiesButton from './ShowDerivedEntitiesButton';
+
+export default ShowDerivedEntitiesButton;
